Deduplicate STT mocks
