+++ tests/components/assist_pipeline/conftest.py
@@ -36,6 +36,7 @@ from tests.common import (
     mock_integration,
     mock_platform,
 )
+from tests.components.stt.common import MockSTTProvider, MockSTTProviderEntity
 
 _TRANSCRIPT = "test transcript"
 
@@ -45,67 +46,6 @@ BYTES_ONE_SECOND = SAMPLE_RATE * SAMPLE_WIDTH * SAMPLE_CHANNELS
 @pytest.fixture(autouse=True)
 def mock_tts_cache_dir_autouse(mock_tts_cache_dir: Path) -> None:
     """Mock the TTS cache dir with empty dir."""
-
-
-class BaseProvider:
-    """Mock STT provider."""
-
-    _supported_languages = ["en-US"]
-
-    def __init__(self, text: str) -> None:
-        """Init test provider."""
-        self.text = text
-        self.received: list[bytes] = []
-
-    @property
-    def supported_languages(self) -> list[str]:
-        """Return a list of supported languages."""
-        return self._supported_languages
-
-    @property
-    def supported_formats(self) -> list[stt.AudioFormats]:
-        """Return a list of supported formats."""
-        return [stt.AudioFormats.WAV]
-
-    @property
-    def supported_codecs(self) -> list[stt.AudioCodecs]:
-        """Return a list of supported codecs."""
-        return [stt.AudioCodecs.PCM]
-
-    @property
-    def supported_bit_rates(self) -> list[stt.AudioBitRates]:
-        """Return a list of supported bitrates."""
-        return [stt.AudioBitRates.BITRATE_16]
-
-    @property
-    def supported_sample_rates(self) -> list[stt.AudioSampleRates]:
-        """Return a list of supported samplerates."""
-        return [stt.AudioSampleRates.SAMPLERATE_16000]
-
-    @property
-    def supported_channels(self) -> list[stt.AudioChannels]:
-        """Return a list of supported channels."""
-        return [stt.AudioChannels.CHANNEL_MONO]
-
-    async def async_process_audio_stream(
-        self, metadata: stt.SpeechMetadata, stream: AsyncIterable[bytes]
-    ) -> stt.SpeechResult:
-        """Process an audio stream."""
-        async for data in stream:
-            if not data:
-                break
-            self.received.append(data)
-        return stt.SpeechResult(self.text, stt.SpeechResultState.SUCCESS)
-
-
-class MockSttProvider(BaseProvider, stt.Provider):
-    """Mock provider."""
-
-
-class MockSttProviderEntity(BaseProvider, stt.SpeechToTextEntity):
-    """Mock provider entity."""
-
-    _attr_name = "Mock STT"
 
 
 class MockTTSProvider(tts.Provider):
@@ -166,15 +106,17 @@ async def mock_tts_provider() -> MockTTSProvider:
 
 
 @pytest.fixture
-async def mock_stt_provider() -> MockSttProvider:
+async def mock_stt_provider() -> MockSTTProvider:
     """Mock STT provider."""
-    return MockSttProvider(_TRANSCRIPT)
+    return MockSTTProvider(supported_languages=["en-US"], text=_TRANSCRIPT)
 
 
 @pytest.fixture
-def mock_stt_provider_entity() -> MockSttProviderEntity:
+def mock_stt_provider_entity() -> MockSTTProviderEntity:
     """Test provider entity fixture."""
-    return MockSttProviderEntity(_TRANSCRIPT)
+    entity = MockSTTProviderEntity(supported_languages=["en-US"], text=_TRANSCRIPT)
+    entity._attr_name = "Mock STT"
+    return entity
 
 
 class MockSttPlatform(MockPlatform):
@@ -290,8 +232,8 @@ def config_flow_fixture(hass: HomeAssistant) -> Generator[None]:
 @pytest.fixture
 async def init_supporting_components(
     hass: HomeAssistant,
-    mock_stt_provider: MockSttProvider,
-    mock_stt_provider_entity: MockSttProviderEntity,
+    mock_stt_provider: MockSTTProvider,
+    mock_stt_provider_entity: MockSTTProviderEntity,
     mock_tts_provider: MockTTSProvider,
     mock_wake_word_provider_entity: MockWakeWordEntity,
     mock_wake_word_provider_entity2: MockWakeWordEntity2,
